move functions to module syntax
