+++ data/db.json
@@ -229,31 +229,6 @@
     },
     {
       "adult": false,
-      "backdrop_path": "/8BTsTfln4jlQrLXUBquXJ0ASQy9.jpg",
-      "genre_ids": [
-        12,
-        28,
-        878
-      ],
-      "id": 140607,
-      "original_language": "en",
-      "original_title": "Star Wars: The Force Awakens",
-      "overview": "Thirty years after defeating the Galactic Empire, Han Solo and his allies face a new threat from the evil Kylo Ren and his army of Stormtroopers.",
-      "popularity": 49.323,
-      "poster_path": "/wqnLdwVXoBjKibFRR5U3y0aDUhs.jpg",
-      "release_date": "2015-12-15",
-      "title": "Star Wars: The Force Awakens",
-      "video": false,
-      "vote_average": 7.295,
-      "vote_count": 18180,
-      "genre_names": [
-        "Adventure",
-        "Action",
-        "Science Fiction"
-      ]
-    },
-    {
-      "adult": false,
       "backdrop_path": "/3MSrtN9M6V8te4XYWeaO2x8EesO.jpg",
       "genre_ids": [
         28,
