+++ movie-project.html
@@ -46,6 +46,6 @@
 <!-- starting of the styling -->
 </div>
 
-<script src="js/movie-project-main.js"></script>
+<script type="module" src="js/movie-project-main.js"></script>
 </body>
 </html>
